server/requirements.txt
bcrypt==4.1.3
fastapi==0.111.0
httpx==0.27.0
python-jose==3.3.0
SQLAlchemy==2.0.31
uvicorn[standard]==0.30.1